LPS-95178 Util functions moved to new file
